LayoutTests/imported/w3c/web-platform-tests/svg/animations/seeking-events-2-expected.txt


FAIL Seeking backwards dispatches 'endEvent' if the element is not active at the seeked-to time assert_equals: Expected beginEvent event, but got endEvent event instead expected "beginEvent" but got "endEvent"
